Use album year as album prefix in 'tracknums' mode
There is no such thing as a track number for an album; but the year is
an adequate substituted.  The advantage here, is that alphabetically
sorting the albums for a particular artist will result in them playing
in chronological order.  For album titles, the alphabetical sorting is
entirely arbitrary, so sort-by-year is far more practical.